Revert "Update unit tests"
This reverts commit 62d4bfdcebdc037f91acf3e31e3624a565e4c555.
