+++ test/unit/pages/ClinicWorkspaceHeader.test.js
@@ -4,6 +4,7 @@ import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import merge from 'lodash/merge';
+import noop from 'lodash/noop';
 import { ToastProvider } from '../../../app/providers/ToastProvider';
 import ClinicWorkspaceHeader from '../../../app/components/clinic/ClinicWorkspaceHeader';
 import Button from '../../../app/components/elements/Button';
@@ -20,13 +21,18 @@ import Button from '../../../app/components/elements/Button';
 const expect = chai.expect;
 const mockStore = configureStore([thunk]);
 
-describe('ClinicWorkspaceHeader', () => {
+describe.only('ClinicWorkspaceHeader', () => {
   let mount;
 
   let wrapper;
   let defaultProps = {
     trackMetric: sinon.stub(),
     t: sinon.stub().callsFake((string) => string),
+    api: {
+      clinics: {
+        update: sinon.stub().callsArgWith(2, null, { updateReturn: 'success' }),
+      },
+    },
   };
 
   before(() => {
@@ -162,7 +168,7 @@ describe('ClinicWorkspaceHeader', () => {
     wrapper = mount(
       <Provider store={store}>
         <ToastProvider>
-          <ClinicWorkspaceHeader {...defaultProps} />
+         .only <ClinicWorkspaceHeader {...defaultProps} />
         </ToastProvider>
       </Provider>
     );
@@ -198,41 +204,7 @@ describe('ClinicWorkspaceHeader', () => {
     wrapper = mount(
       <Provider store={store}>
         <ToastProvider>
-          <ClinicWorkspaceHeader {...defaultProps} />
-        </ToastProvider>
-      </Provider>
-    );
-
-    const link = wrapper.find(Button).filter({ variant: 'textSecondary' });
-    expect(link).to.have.length(1);
-    expect(link.text()).to.equal('View Patient List');
-    expect(link.props().onClick).to.be.a('function');
-    store.clearActions();
-
-    const expectedActions = [
-      {
-        type: '@@router/CALL_HISTORY_METHOD',
-        payload: {
-          args: [
-            '/clinic-workspace',
-          ],
-          method: 'push',
-        },
-      },
-    ];
-
-    link.props().onClick();
-    const actions = store.getActions();
-    expect(actions).to.eql(expectedActions);
-  });
-
-  it('should render a link to the clinic workspace page if currently on clinic profile editing', () => {
-    ClinicWorkspaceHeader.__Rewire__('useLocation', sinon.stub().returns({ pathname: '/clinic-profile'}));
-
-    wrapper = mount(
-      <Provider store={store}>
-        <ToastProvider>
-          <ClinicWorkspaceHeader {...defaultProps} />
+         .only <ClinicWorkspaceHeader {...defaultProps} />
         </ToastProvider>
       </Provider>
     );
@@ -283,19 +255,17 @@ describe('ClinicWorkspaceHeader', () => {
     context('clinic admin team member', () => {
       beforeEach(() => {
         store = mockStore(clinicAdminState);
-        ClinicWorkspaceHeader.__Rewire__('useLocation', sinon.stub().returns({ pathname: '/clinic-workspace'}));
 
         wrapper = mount(
           <Provider store={store}>
             <ToastProvider>
-              <ClinicWorkspaceHeader {...defaultProps} />
+             .only <ClinicWorkspaceHeader {...defaultProps} />
             </ToastProvider>
           </Provider>
         );
       });
 
       it('should show a profile edit button that redirects to the clinic edit form', () => {
-        ClinicWorkspaceHeader.__Rewire__('useLocation', sinon.stub().returns({ pathname: '/clinic-workspace'}));
         const profileButton = wrapper.find('button#profileEditButton');
         expect(profileButton).to.have.lengthOf(1);
         expect(profileButton.text()).to.equal('Edit Clinic Profile');
@@ -317,6 +287,105 @@ describe('ClinicWorkspaceHeader', () => {
         profileButton.simulate('click');
         const actions = store.getActions();
         expect(actions).to.eql(expectedActions);
+      });
+
+      it('should populate the profile edit form with clinic values', () => {
+        const profileButton = wrapper.find('button#profileEditButton');
+        profileButton.simulate('click');
+
+        const profileForm = wrapper.find('form#clinic-profile-update');
+        expect(profileForm).to.have.lengthOf(1);
+
+        expect(profileForm.find('input[name="name"]').prop('value')).to.equal('new_clinic_name');
+        expect(profileForm.find('input[name="phoneNumbers.0.number"]').prop('defaultValue')).to.equal('(888) 555-5555');
+        expect(profileForm.find('select[name="country"]').prop('value')).to.equal('US');
+        expect(profileForm.find('select[name="state"]').prop('value')).to.equal('New Jersey');
+        expect(profileForm.find('input[name="city"]').prop('value')).to.equal('Gotham');
+        expect(profileForm.find('input[name="address"]').prop('value')).to.equal('1 Address Ln, City Zip');
+        expect(profileForm.find('input[name="postalCode"]').prop('value')).to.equal('12345');
+        expect(profileForm.find('input[name="website"]').prop('value')).to.equal('http://clinic.com');
+        expect(profileForm.find('input[name="clinicType"][checked=true]').prop('value')).to.equal('provider_practice');
+        expect(profileForm.find('input[name="clinicSize"][checked=true]').prop('value')).to.equal('0-249');
+        expect(profileForm.find('input[name="preferredBgUnits"][checked=true]').prop('value')).to.equal('mmol/L');
+      });
+
+      it('should submit updated clinic profile values', done => {
+        const profileButton = wrapper.find('button#profileEditButton');
+        profileButton.simulate('click');
+
+        const profileForm = wrapper.find('form#clinic-profile-update');
+        expect(profileForm).to.have.lengthOf(1);
+
+        wrapper.find('input[name="name"]').simulate('change', { persist: noop, target: { name: 'name', value: 'name_updated' } });
+        expect(wrapper.find('input[name="name"]').prop('value')).to.equal('name_updated');
+
+        wrapper.find('input[name="phoneNumbers.0.number"]').simulate('change', { persist: noop, target: { name: 'phoneNumbers.0.number', value: '(888) 555-6666' } });
+        expect(wrapper.find('input[name="phoneNumbers.0.number"]').prop('defaultValue')).to.equal('(888) 555-6666');
+
+        wrapper.find('select[name="country"]').simulate('change', { persist: noop, target: { name: 'country', value: 'CA' } });
+        expect(wrapper.find('select[name="country"]').prop('value')).to.equal('CA');
+
+        wrapper.find('select[name="state"]').simulate('change', { persist: noop, target: { name: 'state', value: 'ON' } });
+        expect(wrapper.find('select[name="state"]').prop('value')).to.equal('ON');
+
+        wrapper.find('input[name="city"]').simulate('change', { persist: noop, target: { name: 'city', value: 'city_updated' } });
+        expect(wrapper.find('input[name="city"]').prop('value')).to.equal('city_updated');
+
+        wrapper.find('input[name="address"]').simulate('change', { persist: noop, target: { name: 'address', value: 'address_updated' } });
+        expect(wrapper.find('input[name="address"]').prop('value')).to.equal('address_updated');
+
+        wrapper.find('input[name="postalCode"]').simulate('change', { persist: noop, target: { name: 'postalCode', value: 'L3X 9G2' } });
+        expect(wrapper.find('input[name="postalCode"]').prop('value')).to.equal('L3X 9G2');
+
+        wrapper.find('input[name="website"]').simulate('change', { persist: noop, target: { name: 'website', value: 'http://clinic_updated.com' } });
+        expect(wrapper.find('input[name="website"]').prop('value')).to.equal('http://clinic_updated.com');
+
+        wrapper.find('input[name="clinicType"]').at(1).simulate('change', { persist: noop, target: { name: 'clinicType', value: 'healthcare_system' } });
+        expect(wrapper.find('input[name="clinicType"][checked=true]').prop('value')).to.equal('healthcare_system');
+
+        wrapper.find('input[name="clinicSize"]').at(1).simulate('change', { persist: noop, target: { name: 'clinicSize', value: '250-499' } });
+        expect(wrapper.find('input[name="clinicSize"][checked=true]').prop('value')).to.equal('250-499');
+
+        wrapper.find('input[name="preferredBgUnits"]').at(1).simulate('change', { persist: noop, target: { name: 'preferredBgUnits', value: 'mg/dL' } });
+        expect(wrapper.find('input[name="preferredBgUnits"][checked=true]').prop('value')).to.equal('mg/dL');
+
+        store.clearActions();
+        wrapper.find('Button#submit').simulate('submit');
+
+        setTimeout(() => {
+          expect(defaultProps.api.clinics.update.callCount).to.equal(1);
+
+          sinon.assert.calledWith(
+            defaultProps.api.clinics.update,
+            'clinicID456',
+            {
+              address: 'address_updated',
+              city: 'city_updated',
+              clinicSize: '250-499',
+              clinicType: 'healthcare_system',
+              country: 'CA',
+              name: 'name_updated',
+              phoneNumbers: [{ number: '(888) 555-6666', type: 'Office' }],
+              postalCode: 'L3X 9G2',
+              state: 'ON',
+              website: 'http://clinic_updated.com',
+              preferredBgUnits: 'mg/dL',
+            }
+          );
+
+          expect(store.getActions()).to.eql([
+            { type: 'UPDATE_CLINIC_REQUEST' },
+            {
+              type: 'UPDATE_CLINIC_SUCCESS',
+              payload: {
+                clinicId: 'clinicID456',
+                clinic: { updateReturn: 'success' },
+              },
+            },
+          ]);
+
+          done();
+        }, 0);
       });
     });
   });
